+++ R/bag_tree-rpart.R
@@ -58,12 +58,18 @@ get_missings_survbagg <- function(object, new_data) {
 #' bagged_tree <- bagging(Surv(time, status) ~ age + ph.ecog, data = lung)
 #' survival_prob_survbagg(bagged_tree, lung[1:3, ], time = 100)
 survival_prob_survbagg <- function(object, new_data, time) {
+  # we could access more than the survival probabilities but
+  # we should not use the standard error and confidence intervals because
+  # "the KM does not know about the tree at all", or more specifically,
+  # how the sample of observations used for the KM is selected based on the tree
+  output <- "surv"
 
+  n_obs <- nrow(new_data)
   missings_in_new_data <- get_missings_survbagg(object, new_data)
+
   if (!is.null(missings_in_new_data)) {
-    n_total <- nrow(new_data)
     n_missing <- length(missings_in_new_data)
-    all_missing <- n_missing == n_total
+    all_missing <- n_missing == n_obs
     if (all_missing) {
       ret <- predict_survival_na(time)
       ret <- tibble(.pred = rep(list(ret), n_missing))
@@ -74,15 +80,23 @@ survival_prob_survbagg <- function(object, new_data, time) {
 
   y <- predict(object, newdata = new_data)
 
-  res <- purrr::map(y, ~ summary(.x, times = pmin(time, max(.x$time)))$surv)
-  res <- matrix(unlist(res), ncol = length(time), byrow = TRUE)
+  survfit_summary_list <- purrr::map(y, summary, times = time, extend = TRUE)
+  survfit_summary_combined <- combine_list_of_survfit_summary(
+    survfit_summary_list,
+    time = time
+  )
 
-  if (!is.null(missings_in_new_data)) {
-    pred_full <- matrix(NA, nrow = n_total, ncol = ncol(res))
-    pred_full[-missings_in_new_data,] <- res
-    res <- pred_full
-  }
+  res <- survfit_summary_patch(
+    survfit_summary_combined,
+    index_missing = missings_in_new_data,
+    time = time,
+    n_obs = n_obs
+  ) %>%
+    survfit_summary_to_tibble(time = time, n_obs = n_obs) %>%
+    keep_cols(output) %>%
+    tidyr::nest(.pred = c(-.row)) %>%
+    dplyr::select(-.row)
 
-  res <- matrix_to_nested_tibbles_survival(res, time)
   res
+
 }